Move generic part of authentication to a Plug
